Add handlers for month and year changes
